Move brandlogo logic to ultility
